libmy: Made major changes to headers, and some minor revamping to code
Changes:
- (outside of libmy) set the entire /include directory as being ignored instead of the specific header files in it, to make it easier to handle
- (both in and outside of libmy) revised Makefile to just let clean remove the obj directory instead of doing it twice
- Dummied out my.h (i.e. now it just errors and redirects you to the new header files)
- Straight up removed my_bigint.h, my_bigint_internal.h and my_config.h
- Added all these header files and moved relevant parts of my.h to them :
my/bigint.h
my/checked_multiply.h
my/config.h
my/ctype.h
my/internal/bigint.h
my/internal/getnbr_part2.h
my/macros.h
my/math.h
my/misc.h
my/stdio.h
my/stdlib.h
my/string.h
- (also outside of libmy) Corrected source files to use appropriate headers instead of my.h, also took the occasion to remove some redundant headers from some source files that included too much
- Moved my_checked_multiply to my_checked_multiply_int and added my_checked_multiply_unsigned and modified source files to go along with this change
- Changed my_is_prime to take unsigned instead of int and return bool
- Removed my_sort_int_array as it is made useless by my_qsort
- Made my_strn{cat,cpy} take a size_t instead of an int for the last argument, for consistency with libc

+++ lib/my/src/my_strncpy.c
@@ -8,14 +8,14 @@
 ** additional null bytes to dst to ensure a total of n bytes are written
 */
 
-#include "my.h"
+#include "my/string.h"
 
-char *my_strncpy(char *dest, const char *src, int n)
+char *my_strncpy(char *dest, const char *src, size_t n)
 {
     char *result = dest;
     char current_character;
 
-    for (int i = 0; i < n; ++i) {
+    for (size_t i = 0; i < n; ++i) {
         current_character = *src++;
         *dest++ = current_character;
